+++ src/main/java/nl/edemtb/mtbclinicsapplication/mappers/RouteMapper.java
@@ -1,7 +1,7 @@
 package nl.edemtb.mtbclinicsapplication.mappers;
 
-import nl.edemtb.mtbclinicsapplication.dtos.RouteDto;
-import nl.edemtb.mtbclinicsapplication.dtos.RouteInputDto;
+import nl.edemtb.mtbclinicsapplication.dtos.route.RouteDto;
+import nl.edemtb.mtbclinicsapplication.dtos.route.RouteInputDto;
 import nl.edemtb.mtbclinicsapplication.models.Route;
 import nl.edemtb.mtbclinicsapplication.repositories.RouteRepository;
 import org.springframework.stereotype.Component;
